Block drivers: make IOCTL request unsigned long
The block driver protocol and libblockdriver's bdr_ioctl hook are
changed, as well as the users of this hook. Other parts of the system
are expected to change accordingly eventually, since the ioctl(2)
prototype has been aligned with NetBSD's.

+++ include/minix/blockdriver.h
@@ -19,7 +19,7 @@ struct blockdriver {
   int (*bdr_close)(devminor_t minor);
   ssize_t (*bdr_transfer)(devminor_t minor, int do_write, u64_t pos,
 	  endpoint_t endpt, iovec_t *iov, unsigned int count, int flags);
-  int (*bdr_ioctl)(devminor_t minor, unsigned int request, endpoint_t endpt,
+  int (*bdr_ioctl)(devminor_t minor, unsigned long request, endpoint_t endpt,
 	  cp_grant_id_t grant, endpoint_t user_endpt);
   void (*bdr_cleanup)(void);
   struct device *(*bdr_part)(devminor_t minor);
